Resolve circular imports for `memos`
Specifically:

* `callbacks` require `containers` (now: memos) for the kwarg `memo`.
* `containers` require `handlers` for `HandlerId` & daemon handler
* `handlers` require `callbacks` for callback protocols

Previously, it was not a problem because `callbacks` did not declare `memo` kwargs by mistake, and therefore did not import `containers`.

The chain is broken at the `callbacks`→`containers` import: it is now `callbacks`→`memos`.
