+++ kopf/structs/containers.py
@@ -10,9 +10,9 @@ object, even if that object does not show up in the event streams for long time.
 import dataclasses
 import logging
 import time
-from typing import Any, Dict, Iterator, MutableMapping, Optional, Set, Union
+from typing import Dict, Iterator, MutableMapping, Optional, Set, Union
 
-from kopf.structs import bodies, handlers, primitives
+from kopf.structs import bodies, handlers, memos, primitives
 from kopf.utilities import aiotasks
 
 
@@ -32,31 +32,12 @@ class Throttler:
     active_until: Optional[float] = None  # internal clock
 
 
-class Memo(Dict[Any, Any]):
-    """ A container to hold arbitrary keys-fields assigned by the users. """
-
-    def __setattr__(self, key: str, value: Any) -> None:
-        self[key] = value
-
-    def __delattr__(self, key: str) -> None:
-        try:
-            del self[key]
-        except KeyError as e:
-            raise AttributeError(str(e))
-
-    def __getattr__(self, key: str) -> Any:
-        try:
-            return self[key]
-        except KeyError as e:
-            raise AttributeError(str(e))
-
-
 @dataclasses.dataclass(frozen=False)
 class ResourceMemory:
     """ A system memo about a single resource/object. Usually stored in `Memories`. """
 
     # For arbitrary user data to be stored in memory, passed as `memo` to all the handlers.
-    memo: Memo = dataclasses.field(default_factory=Memo)
+    memo: memos.Memo = dataclasses.field(default_factory=memos.Memo)
 
     # For resuming handlers tracking and deciding on should they be called or not.
     noticed_by_listing: bool = False
