+++ tests/registries/conftest.py
@@ -8,9 +8,9 @@ from kopf.reactor.registries import ActivityRegistry, OperatorRegistry, Resource
                                     ResourceRegistry, ResourceSpawningRegistry, \
                                     ResourceWatchingRegistry
 from kopf.structs.bodies import Body
-from kopf.structs.containers import Memo
 from kopf.structs.diffs import Diff, DiffItem
 from kopf.structs.handlers import HandlerId, ResourceChangingHandler
+from kopf.structs.memos import Memo
 from kopf.structs.patches import Patch
